Move model SPI interfaces from maven-impl to maven-api-spi
Move 23 interfaces from the org.apache.maven.api.services.model
package in impl/maven-impl to api/maven-api-spi where they
belong as part of the public API surface.

Add proper Javadoc to all moved interfaces.
