+++ api/maven-api-spi/src/main/java/org/apache/maven/api/services/model/DependencyManagementImporter.java
@@ -18,28 +18,31 @@
  */
 package org.apache.maven.api.services.model;
 
-import java.util.Collection;
 import java.util.List;
 
-import org.apache.maven.api.model.Profile;
+import org.apache.maven.api.model.DependencyManagement;
+import org.apache.maven.api.model.Model;
+import org.apache.maven.api.services.ModelBuilderRequest;
 import org.apache.maven.api.services.ModelProblemCollector;
 
 /**
- * Calculates the active profiles among a given collection of profiles.
+ * Handles the import of dependency management from other models into the target model.
  *
+ * @since 4.0.0
  */
-public interface ProfileSelector {
+public interface DependencyManagementImporter {
 
     /**
-     * Determines the profiles which are active in the specified activation context. Active profiles will eventually be
-     * injected into the model.
+     * Imports the specified dependency management sections into the given target model.
      *
-     * @param profiles The profiles whose activation status should be determined, must not be {@code null}.
-     * @param context The environmental context used to determine the activation status of a profile, must not be
-     *            {@code null}.
+     * @param target The model into which to import the dependency management section, must not be <code>null</code>.
+     * @param sources The dependency management sections to import, may be <code>null</code>.
+     * @param request The model building request that holds further settings, must not be {@code null}.
      * @param problems The container used to collect problems that were encountered, must not be {@code null}.
-     * @return The profiles that have been activated, never {@code null}.
      */
-    List<Profile> getActiveProfiles(
-            Collection<Profile> profiles, ProfileActivationContext context, ModelProblemCollector problems);
+    Model importManagement(
+            Model target,
+            List<? extends DependencyManagement> sources,
+            ModelBuilderRequest request,
+            ModelProblemCollector problems);
 }
