+++ api/maven-api-spi/src/main/java/org/apache/maven/api/services/model/ModelValidator.java
@@ -25,6 +25,7 @@ import org.apache.maven.api.services.ModelProblemCollector;
 /**
  * Checks the model for missing or invalid values.
  *
+ * @since 4.0.0
  */
 public interface ModelValidator {
     /**
