+++ api/maven-api-spi/src/main/java/org/apache/maven/api/services/model/ProfileInjector.java
@@ -28,6 +28,7 @@ import org.apache.maven.api.services.ModelProblemCollector;
 /**
  * Handles profile injection into the model.
  *
+ * @since 4.0.0
  */
 public interface ProfileInjector {
